HBN-650 change decimals declaration to ERC20 Standard of uint8

+++ contracts/ExternStateToken.sol
@@ -50,7 +50,7 @@ contract ExternStateToken is SelfDestructible, Proxyable, TokenFallbackCaller {
     string public name;
     string public symbol;
     uint public totalSupply;
-    uint public decimals;
+    uint8 public decimals;
 
     /**
      * @dev Constructor.
